+++ api/package.json
@@ -13,16 +13,14 @@
     "nodemon": "^2.0.15"
   },
   "dependencies": {
-
-    "dotenv": "^10.0.0",
     "bcrypt": "^5.0.1",
+    "dotenv": "^8.2.0",
     "express": "^4.17.2",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "passport": "^0.5.2",
     "passport-jwt": "^4.0.0",
     "pg": "^8.7.1",
-    "sequelize": "^6.12.5",
-    "dotenv": "^8.2.0"
+    "sequelize": "^6.12.5"
   }
 }
