--- api/src/models/proveedor.js
@@ -1,37 +1,0 @@
-const { DataTypes } = require('sequelize');
-// Exportamos una funcion que define el modelo
-// Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
-  // defino el modelo
-  sequelize.define('proveedor', {
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    lastname: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    mail: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    access: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    cuit: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    businessname: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-
-    })
-  }
