+++ api/src/routes/user/controllers/register.js
@@ -1,13 +1,13 @@
-const User = require("../../models/User")
+const { User } = require("../../../db")
 const bcrypt = require("bcrypt")
 
-export const register = async (req, res) => {
-
+const register = async (req, res) => {
+    console.log(req.body)
     const { mail, password, /* etc */ } = req.body
-
+    
     const encryptedPassword = bcrypt.hashSync(password, 10)
 
-    const alreadyExists = await User.findOne({ where: { mail }}).catch(err => console.log(err))
+    const alreadyExists = await User.findOne({ where: { mail: mail }}).catch(err => console.log(err))
 
     if(alreadyExists) return res.json({ message: "Ya existe un usuario con ese email" })
     
@@ -20,4 +20,8 @@ export const register = async (req, res) => {
         res.json({ error: "No se puede registrar al usuario en este momento" })
     })
     if (newUser) res.json({ message: "El usuario se ha registrado correctamente" })
+}
+
+module.exports = {
+    register
 }
